+++ src/daffodil/lib/daf_pandas.py
@@ -69,7 +69,7 @@ def _from_pandas_df(
         name: str='', 
         use_csv: bool=False, 
         dtypes: Optional[T_dtype_dict]=None
-        ):
+        ):  # -> 'Daf'
     """
     Convert a Pandas dataframe to daf object
         @@TODO: This does not enforce dtypes are correct.